requirements.txt
# ============================================================================
# DEPENDENCIAS PYTHON - PUNTAFINA ETL BATCH
# ============================================================================
# Sistema ETL optimizado para procesamiento por lotes en Ubuntu 22.04

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0

# Data formats
pyarrow>=12.0.0
fastparquet>=2023.0.0

# System monitoring
psutil>=5.9.0

# CLI and UI
click>=8.1.0
tqdm>=4.65.0
tabulate>=0.9.0
colorama>=0.4.6

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0

# Logging and monitoring
loguru>=0.7.0

# Data validation
cerberus>=1.3.0
jsonschema>=4.17.0

# Performance
numba>=0.57.0
